refactor(config): comprehensive security and validation improvements
🔒 SECURITY FIXES:
- Fix temp files sanitizer to prevent current directory exclusion
- Ensure duration validation runs regardless of TrimWhitespace flag
- Change Nix generations from silent fix to explicit validation error
- Extend Nix generations range from 1-10 to 1-1000 with proper error handling

🛡️ VALIDATION IMPROVEMENTS:
- Replace direct type assertions with errors.As for wrapped errors
- Add safe always-false regex fallback in GetCompiledRegex to prevent panics
- Implement centralized critical risk + safe mode validation helper
- Fix profile field consistency to use map keys instead of .Name property
- Add nil profile guards to prevent panics in validation loops

🔧 TECHNICAL REFACTORING:
- Consolidate duplicate critical risk validation into single helper function
- Remove duplicate safe mode warnings in favor of per-operation errors
- Fix DI container type assertions and interface design
- Improve test error handling and type safety

📋 BUSINESS LOGIC:
- Proper error propagation in sanitization pipeline
- Enhanced validation result consistency across modules
- Improved field path naming conventions throughout validation

This refactoring significantly improves the security posture and reliability
of the configuration validation system while maintaining backward compatibility.

+++ cmd/clean-wizard/commands/clean.go
@@ -202,7 +202,7 @@ func displayCleanResults(result domain.CleanResult, verbose bool, duration time.
 	if result.IsValid() {
 		fmt.Printf("   • Status: %d items %s\n", result.ItemsRemoved, action)
 		if result.FreedBytes > 0 {
-			fmt.Printf("   • Space freed: %s\n", format.Bytes(result.FreedBytes))
+			fmt.Printf("   • Space freed: %s\n", format.Bytes(int64(result.FreedBytes)))
 		}
 
 		if verbose {
